extract `lib.user.mail`
